feat: add day_CH to Category table & modify time format by moment & add time and day to flash message

+++ controller/enrollController.js
@@ -1,14 +1,14 @@
 const db = require('../models')
-const { Enroll, Trainingday, WaitingList } = db
+const { Enroll, Trainingday, WaitingList, Category } = db
 const moment = require('moment')
 
 const enrollController = {
   enrollCourse: async (req, res) => {
     const limitNumbers = 1 // 每堂課最多20人
     // 報名的課程日期
-    const trainingDay = await Trainingday.findByPk(req.params.id, { raw: true })
+    const trainingDay = await Trainingday.findByPk(req.params.id, { raw: true, nest: true, include: [Category] })
     trainingDay.date = moment(trainingDay.date).format('YYYY-MM-DD')
-
+    trainingDay.time = moment(trainingDay.time, moment.HTML5_FMT.TIME).format("HH:mm")
     // // 確認是否已報名過
     // const alreadyEnroll = await Enroll.findOne({ where: { UserId: req.user.id, TrainingdayId: req.params.id } })
     // if (alreadyEnroll) {
@@ -29,7 +29,7 @@ const enrollController = {
     // 目前該堂正取人數尚未額滿
     if (enrollCount < limitNumbers) {
       await Enroll.create({ UserId: Number(req.user.id), TrainingdayId: Number(req.params.id) }, { raw: true })
-      req.flash('success_msg', `已成功${trainingDay.date}報名`)
+      req.flash('success_msg', `已成功 ${trainingDay.date} (${trainingDay.Category.day_CH}) ${trainingDay.time} 報名`)
       return res.redirect('/user/training')
     }
 
@@ -38,7 +38,7 @@ const enrollController = {
       // 確認備取名單尚未額滿
       if (waitingCount < limitNumbers) {
         await WaitingList.create({ UserId: Number(req.user.id), TrainingdayId: Number(req.params.id) }, { raw: true })
-        req.flash('success_msg', `已備取${trainingDay.date}課程`)
+        req.flash('success_msg', `已備取 ${trainingDay.date} (${trainingDay.Category.day_CH}) ${trainingDay.time} 課程`)
         return res.redirect('/user/training')
       }
     }
@@ -46,9 +46,7 @@ const enrollController = {
 
     // 報名額滿，備取已額滿
     if (enrollCount >= limitNumbers && waitingCount >= limitNumbers) {
-      console.log(enrollCount)
-      console.log(waitingCount)
-      req.flash('warning_msg', `${trainingDay.date}報名已額滿`)
+      req.flash('warning_msg', `${trainingDay.date} (${trainingDay.Category.day_CH}) ${trainingDay.time} 報名已額滿`)
       return res.redirect('/user/training')
     }
 
@@ -56,21 +54,29 @@ const enrollController = {
   cancelEnroll: async (req, res) => {
     // 取消正取
     await Enroll.destroy({ where: { UserId: Number(req.user.id), TrainingdayId: Number(req.params.id) } })
-    const cancel = await Trainingday.findByPk(req.params.id)
+    const cancel = await Trainingday.findByPk(req.params.id, { raw: true, nest: true, include: [Category] })
+    cancel.date = moment(cancel.date).format('YYYY-MM-DD')
+    cancel.time = moment(cancel.time, moment.HTML5_FMT.TIME).format("HH:mm")
 
-    //目前備取第一位要改為正取
-    const { UserId, TrainingdayId } = await WaitingList.findOne({ where: { TrainingdayId: Number(req.params.id) }, order: [['createdAt', 'ASC']], limit: 1, include: [Trainingday] })
-    await WaitingList.destroy({ where: { UserId, TrainingdayId } })
-    await Enroll.create({ UserId, TrainingdayId })
+    // 如果目前有備取名單
+    // 目前備取第一位要改為正取
+    const onWaiting = await WaitingList.findOne({ where: { TrainingdayId: Number(req.params.id) }, order: [['createdAt', 'ASC']], limit: 1, include: [Trainingday] })
+    if (onWaiting) {
+      const { UserId, TrainingdayId } = onWaiting
+      await WaitingList.destroy({ where: { UserId, TrainingdayId } })
+      await Enroll.create({ UserId, TrainingdayId })
+    }
 
-    req.flash('success_msg', `已取消報名 ${moment(cancel.date).format('YYYY-MM-DD')} 課程`)
+    req.flash('success_msg', `已取消報名 ${cancel.date} (${cancel.Category.day_CH}) ${cancel.time} 課程`)
     return res.redirect('/user/training')
   },
   cancelWaiting: async (req, res) => {
     // 取消備取
     await WaitingList.destroy({ where: { UserId: Number(req.user.id), TrainingdayId: Number(req.params.id) } })
-    const cancel = await Trainingday.findByPk(req.params.id)
-    req.flash('success_msg', `已取消報名 ${moment(cancel.date).format('YYYY-MM-DD')} 課程`)
+    const cancel = await Trainingday.findByPk(req.params.id, { raw: true, nest: true, include: [Category] })
+    cancel.date = moment(cancel.date).format('YYYY-MM-DD')
+    cancel.time = moment(cancel.time, moment.HTML5_FMT.TIME).format("HH:mm")
+    req.flash('success_msg', `已取消報名 ${cancel.date} (${cancel.Category.day_CH}) ${cancel.time} 課程`)
     return res.redirect('/user/training')
   }
